3-factorial.c

+++ recursion/3-factorial.c
@@ -1,22 +1,20 @@
 #include <stdio.h>
 
-unsigned long long factorial(int n);
+unsigned long factorial(int n);
 
 int main(void)
 {
     int number;
     
-   
     printf("Enter a number: ");
     scanf("%d", &number);
 
-    
-    printf("Factorial of %d = %llu\n", number, factorial(number));
+    printf("Factorial of %d = %lu\n", number, factorial(number));
 
     return 0;
 }
 
-unsigned long long factorial(int n)
+unsigned long factorial(int n)
 {
     if (n == 0 || n == 1)
         return 1;
